meta_ai/requirements.txt
openai==0.28.1
pymongo
python-dateutil
PyYAML
